requirements.txt
ale-py==0.7.4
einops
gym[accept-rom-license]
hydra-core==1.1.1
opencv-python
protobuf==3.20.*
psutil
pygame
tqdm
wandb
transformers
numpy
pillow
torchvision
dataclasses
requests
matplotlib
omegaconf
yaml
pyyaml